json/locale: restore the locale after parsing floatingpoint numbers
The parsing of floating points depends on the locale. We need to use
a fixed one in order to be interoperable, but the calling application
is free to set whatever locale they wish.

For this reason we change the locale of the current thread to "C"
while parsing floating point numbers. This patch makes sure to
change back to the previous locale before returning from the function.

Moreover, rather than generating the "C" locale on every call, cache it
in the CJson object.

Signed-off-by: Tom Gundersen <teg@jklm.no>